Ldap schema structure cleanup
