+++ src/Schemas/OpenLDAP.php
@@ -2,8 +2,8 @@
 
 namespace Adldap\Schemas;
 
-class OpenLDAP extends ActiveDirectory
-{
+class OpenLDAP extends BaseLdap {
+
     /**
      * {@inheritdoc}
      */
